31 may

+++ ft_strcat.c
@@ -6,16 +6,16 @@
 /*   By: smadesi <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/21 10:06:27 by smadesi           #+#    #+#             */
-/*   Updated: 2019/05/22 11:19:53 by smadesi          ###   ########.fr       */
+/*   Updated: 2019/05/31 11:50:27 by smadesi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <string.h>
+#include "libft.h"
 
 char	*ft_strcat(char *s1, const char *s2)
 {
-	int i;
-	int j;
+	size_t i;
+	size_t j;
 
 	i = 0;
 	j = 0;
